ResumAI/requirements.txt
streamlit==1.46.0
streamlit-extras
PyPDF2
requests
python-dotenv
